Revert "relocate SlotUtil"
This reverts commit d34824e4c30ae982f828a2775f336999bbde1b6c.
